+++ NEWS.md
@@ -1,5 +1,9 @@
 # readr 0.2.2.9000
 
+* Fix bug in `read_fwf()`, it will now properly read a subset of columns.
+  If the final column is ragged, supply an NA as the final end `fwf_positions`
+  or final width `fwf_widths` position (#353,@ghaarsma).
+
 * readr now imports tibble so that you get consistent `tbl_df` behaviour 
   (#317, #385).